requirements.txt
streamlit
requests
beautifulsoup4
langchain
langchain-community
langchain-core
transformers
torch
faiss-cpu
python-dotenv
